组织列表获取，获取组织分享码

+++ SoundMentor-web/src/main/java/com/soundmentor/soundmentorweb/service/UserInfoApi.java
@@ -14,6 +14,6 @@ public interface UserInfoApi {
     void setUser(UserDO user);
     UserDO getUser();
     List<OrganizationDO> getOrganizations();
-    OrganizationRole getOrganizationRole(Long organizationId);
+    OrganizationRole getOrganizationRole(Integer organizationId);
     void removeUser();
 }
